7811: Move display_mismatch_policy field to a dedicated base class
display_mismatch_policy is only honoured by validators that perform
\$validate-code semantics (in-memory and remote terminology). Hoisting it onto
BaseValidationSupport — as the earlier commit on this branch did — leaked the
setting onto unrelated subclasses (PrePopulatedValidationSupport,
CachingValidationSupport, BaseStaticResourceValidationSupport,
UnknownCodeSystemWarningValidationSupport, …) where it has no effect.

Introduce BaseTerminologyServerValidationSupport between BaseValidationSupport
and the two terminology validators; move the field and its accessors there.
BaseValidationSupport returns to just FhirContext + getFhirContext().
External callers (e.g. ValidationSupportConfig wiring) hold the concrete
subclass type and inherit the setter through the new intermediate class.
